requirements.txt
PyQt6
qtawesome
Pillow
pydub
ffmpeg-python
python-docx
reportlab
